Fs split update 2 (submodules and merge conflicts fixes) (#1)
* Implemented AMR cleanup which sprung from GPU branch. Still has linking error.

* Fixed compilation but still some difficulties

* Updated gitignore

* miscellaneous small fixes. Single-task translation works fine, multi-task gets too much accumulated f into some cells, causing dt to plummet.

* extra debug outputs

* Found reason for weird single-task f accumulation. MPI multi-task diffs still grow.

* Extra printouts

* Finally found the issue of f growth; in target cells, the ratio was mistakenly converted to int, preventing 0.25 ratio cells from getting reset.

* Compilation fix and use newer-style auto members in face neighbours

* simplify seed id selection over periodic dimensions

* get rid of compiler warnigns

* FSgrid commit update

* Timer cleanup

* Responses to Yann's PR review

* Don't close directories that aren't opened, improve syntax

* Todos are not warnings

* Remove apparently orphaned define

* Ensure abort_mpi does not return

* Ensure that abort_mpi actually aborts even if it's not called by all processes

* Unrefine all cells outside refine_radius

* Improve neighbor refinement evaluation

* Fix parens

* Reverting submodule urls to https instead of ssh, for easier accessibility. These are not supposed to be pushed to directly anyway.

* Face neighbor caching submodule

* Use neighbor radius

* Cross-compile on BSC RiscV cluster for much faster CI builds

* Squelch a big bunch of unused variable warnings.

Mostly species parameters in semi-abandoned project definitions, that
have never been used for anything.

* dummy

* dummy

* dummy

* Multipop optimization in sysboundaries: only calculate new moments once

* Also set population-specific moments to zero in case of no blocks

* Added comment about FS-SPLIT and LBWEIGHTCOUNTER

---------
