+++ .gitmodules
@@ -1,12 +1,11 @@
 [submodule "submodules/fsgrid"]
 	path = submodules/fsgrid
-	url = git@github.com:hokkanen/fsgrid.git
-	branch = fs-split
+	url = https://github.com/fmihpc/fsgrid.git
 
 [submodule "submodules/dccrg"]
 	path = submodules/dccrg
-	url = git@github.com:hokkanen/dccrg.git
-	branch = dccrg-split
+	url = https://github.com/fmihpc/dccrg.git
+	branch = vlasiator-version
 
 [submodule "submodules/vectorclass"]
 	path = submodules/vectorclass
