+++ MAKE/Makefile.arriesgado
@@ -39,21 +39,21 @@ LIB_MPI =
 
 LIBRARY_PREFIX = libraries-arriesgado
 
-INC_BOOST =
-LIB_BOOST = ${BOOST_LIBS} -lboost_program_options
+INC_BOOST = -I/apps/riscv/boost/1.77.0/include
+LIB_BOOST = -L/apps/riscv/boost/1.77.0/lib -lboost_program_options
 
 INC_ZOLTAN = -I${LIBRARY_PREFIX}/include
 LIB_ZOLTAN = -L${LIBRARY_PREFIX}/lib -lzoltan
 
-INC_VLSV = -I$(LIBRARY_PREFIX)/vlsv
-LIB_VLSV = -L$(LIBRARY_PREFIX)/vlsv -lvlsv
+INC_VLSV = -I$(LIBRARY_PREFIX)/include
+LIB_VLSV = -L$(LIBRARY_PREFIX)/lib -lvlsv
 
 INC_DCCRG = -I./submodules/dccrg
 
 INC_FSGRID = -I./submodules/fsgrid
 
-LIB_PROFILE = -L${LIBRARY_PREFIX}/phiprof/lib -lphiprof -Wl,-rpath=${LIBRARY_PREFIX}/lib 
-INC_PROFILE = -I${LIBRARY_PREFIX}/phiprof/include
+LIB_PROFILE = -L${LIBRARY_PREFIX}/lib -lphiprof -Wl,-rpath=${LIBRARY_PREFIX}/lib 
+INC_PROFILE = -I${LIBRARY_PREFIX}/include
 INC_TOPO =
 
 INC_EIGEN = ${EIGEN3_INCL}
