+++ projects/Riemann1/Riemann1.cpp
@@ -117,8 +117,8 @@ namespace projects {
                   const std::array<Real, 3> xyz = perBGrid.getPhysicalCoords(x, y, z);
                   std::array<Real, fsgrids::bfield::N_BFIELD>* cell = perBGrid.get(x, y, z);
                   
-                  Real Bxavg, Byavg, Bzavg;
-                  Bxavg = Byavg = Bzavg = 0.0;
+                  //Real Bxavg, Byavg, Bzavg;
+                  //Bxavg = Byavg = Bzavg = 0.0;
                   cell->at(fsgrids::bfield::PERBX) = (xyz[0] < 0.0) ? this->Bx[this->LEFT] : this->Bx[this->RIGHT];
                   cell->at(fsgrids::bfield::PERBY) = (xyz[0] < 0.0) ? this->By[this->LEFT] : this->By[this->RIGHT];
                   cell->at(fsgrids::bfield::PERBZ) = (xyz[0] < 0.0) ? this->Bz[this->LEFT] : this->Bz[this->RIGHT];
