+++ sysboundary/ionosphere.cpp
@@ -882,7 +882,6 @@ namespace SBC {
       if(!refillTensorAtRestart) {
          // Ranks that don't participate in ionosphere solving skip this function outright
          if(!isCouplingInwards && !isCouplingOutwards) {
-            phiprof::stop("ionosphere-calculateConductivityTensor");
             return;
          }
